requirements.txt
Flask
flask-cors
ultralytics
opencv-python
cloudinary
firebase-admin
python-dotenv
supabase